Fix one test in context.rst on Plone 5.1 by being less strict.
On 5.1 instead of `text/html; charset=utf-8` we get `text/html; charset=ASCII`.
This is caused by code that I added in this `context.rst` file to make the tests pass after a Zope security fix that landed in Plone 5.2.10/11.
This test fix is not needed on Plone 5.1.
We only care that we get text/html, and not text/plain, and that part of the test works on all Plone versions.